chore/chain Add group_size requirement
Use group_size to allow blocks to have majority or group_size as required signatories. This allows splits to be caered for as a group will grow until it splits and then lose poss 50% of group, denying our ability to get a majority. This chancge allows group_size to be the required signatories in any case. Group size should not grow to group_size * 2 in a normal working network in any case.